Automatisation/Installation llavax4 - Docker/web/app/requirements.txt
torch
transformers
accelerate
sentencepiece
torchvision
fastapi
uvicorn
pillow
git+https://github.com/huggingface/peft
git+https://github.com/haotian-liu/LLaVA.git